Moving key auth checks into ServiceManager class. Ensuring that settings errors are being displayed
